backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
databases[postgresql]==0.8.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
